# Conflicts:
#	Helpers/Graphics/GL/GLHelper.c
#	main.c

+++ Assets/shader/GL/fb_blur_f.glsl
@@ -10,16 +10,18 @@ out vec4 COLOR;
 
 // Uniforms
 uniform sampler2D alb;
-uniform float blurRadius; // Radius of the blur
+uniform int blurRadius; // Radius of the blur
 
 void main() {
     vec2 uv = gl_FragCoord.xy / textureSize(alb, 0);
+    //COLOR = textureLod(alb, uv, 4.0);
+    //return;
 
     // Box blur
     vec3 color = vec3(0.0);
     ivec2 tex_size = textureSize(alb, 0);
-    for (int x = -int(blurRadius); x <= int(blurRadius); x++) {
-        for (int y = -int(blurRadius); y <= int(blurRadius); y++) {
+    for (int x = -blurRadius; x <= blurRadius; x++) {
+        for (int y = -blurRadius; y <= blurRadius; y++) {
             color += texture(alb, uv + vec2(x, y) / tex_size).rgb;
         }
     }
